+++ src/BTrees/Interfaces.py
@@ -13,35 +13,29 @@
 ##############################################################################
 
 from zope.interface import Interface, Attribute
+from zope.interface.common.collections import ISized
+from zope.interface.common.sequence import IMinimalSequence
+from zope.interface.common.collections import IMapping
 
+# pylint:disable=inherit-non-class,no-method-argument,no-self-argument
+# pylint:disable=unexpected-special-method-signature
 
 class ICollection(Interface):
+    """
+    A collection of zero or more objects.
+
+    In a boolean context, objects implementing this interface are
+    `True` if the collection is non-empty, and `False` if the
+    collection is empty.
+    """
 
     def clear():
         """Remove all of the items from the collection."""
 
-    def __nonzero__():
-        """Check if the collection is non-empty.
 
-        Return a true value if the collection is non-empty and a
-        false value otherwise.
-        """
-
-
-class IReadSequence(Interface):
-
-    def __getitem__(index):
-        """Return the value at the given index.
-
-        An :class:`IndexError` is raised if the index cannot be found.
-        """
-
-    def __getslice__(index1, index2):
-        """Return a subsequence from the original sequence.
-
-        The subsequence includes the items from index1 up to, but not
-        including, index2.
-        """
+# Backwards compatibility alias. To be removed in 5.0.
+# Docs deprecated only in docs/api.rst.
+IReadSequence = IMinimalSequence
 
 class IKeyed(ICollection):
 
@@ -52,25 +46,29 @@ class IKeyed(ICollection):
         """
 
     def keys(min=None, max=None, excludemin=False, excludemax=False):
-        """Return an :class:`~BTrees.Interfaces.IReadSequence` containing the keys in the collection.
+        """
+        Return an :mod:`IMinimalSequence <zope.interface.common.sequence>`
+        containing the keys in the collection.
 
-        The type of the :class:`~BTrees.Interfaces.IReadSequence` is not specified.  It could be a list
-        or a tuple or some other type.
+        The type of the ``IMinimalSequence`` is not specified. It
+        could be a `list` or a `tuple` or some other type.
 
         All arguments are optional, and may be specified as keyword
         arguments, or by position.
 
-        If a min is specified, then output is constrained to keys greater
-        than or equal to the given min, and, if excludemin is specified and
-        true, is further constrained to keys strictly greater than min.  A
-        min value of None is ignored. If min is None or not specified, and
-        excludemin is true, the smallest key is excluded.
+        If a *min* is specified, then output is constrained to keys
+        greater than or equal to the given min, and, if *excludemin*
+        is specified and true, is further constrained to keys strictly
+        greater than *min*. A *min* value of `None` is ignored. If
+        *min* is `None` or not specified, and *excludemin* is true,
+        the smallest key is excluded.
 
-        If a max is specified, then output is constrained to keys less than
-        or equal to the given max, and, if excludemax is specified and
-        true, is further constrained to keys strictly less than max.  A max
-        value of None is ignored.  If max is None or not specified, and
-        excludemax is true, the largest key is excluded.
+        If a *max* is specified, then output is constrained to keys
+        less than or equal to the given *max*, and, if *excludemax* is
+        specified and true, is further constrained to keys strictly
+        less than *max*. A *max* value of `None` is ignored. If *max*
+        is `None` or not specified, and *excludemax* is true, the
+        largest key is excluded.
         """
 
     def maxKey(key=None):
@@ -108,13 +106,6 @@ class ISetMutable(IKeyed):
         """Add the items from the given sequence to the set."""
 
 
-class ISized(Interface):
-    """An object that supports __len__."""
-
-    def __len__():
-        """Return the number of items in the container."""
-
-
 class IKeySequence(IKeyed, ISized):
 
     def __getitem__(index):
@@ -125,7 +116,7 @@ class IKeySequence(IKeyed, ISized):
         """
 
 
-class ISet(IKeySequence, ISetMutable):
+class ISet(ISetMutable, IKeySequence):
     def __and__(other):
         """Shortcut for :meth:`~BTrees.Interfaces.IMerge.intersection`"""
 
@@ -147,7 +138,14 @@ class ITreeSet(ISetMutable):
         """Shortcut for :meth:`~BTrees.Interfaces.IMerge.difference"""
 
 
-class IMinimalDictionary(ISized, IKeyed):
+
+class IMinimalDictionary(IKeyed, IMapping):
+    """
+    Mapping operations.
+
+    .. versionchanged:: 4.8.0
+       Now extends :class:`zope.interface.common.collections.IMapping`.
+    """
 
     def get(key, default):
         """Get the value associated with the given key.
@@ -174,53 +172,61 @@ class IMinimalDictionary(ISized, IKeyed):
         """
 
     def values(min=None, max=None, excludemin=False, excludemax=False):
-        """Return an :class:`BTrees.Interfaces.IReadSequence` containing the values in the collection.
+        """
+        Return an :mod:`IMinimalSequence <zope.interface.common.sequence.IMinimalSequence>`
+        containing the values in the collection.
 
-        The type of the :class:`~BTrees.Interfaces.IReadSequence` is not specified. It could be a list
-        or a tuple or some other type.
+        The type of the ``IMinimalSequence`` is not specified. It
+        could be a `list` or a `tuple` or some other type.
 
         All arguments are optional, and may be specified as keyword
         arguments, or by position.
 
-        If a min is specified, then output is constrained to values whose
-        keys are greater than or equal to the given min, and, if excludemin
-        is specified and true, is further constrained to values whose keys
-        are strictly greater than min.  A min value of None is ignored.  If
-        min is None or not specified, and excludemin is true, the value
-        corresponding to the smallest key is excluded.
+        If a *min* is specified, then output is constrained to values
+        whose keys are greater than or equal to the given *min*, and, if
+        *excludemin* is specified and true, is further constrained to
+        values whose keys are strictly greater than *min*. A *min* value
+        of `None` is ignored. If *min* is `None` or not specified, and
+        *excludemin* is true, the value corresponding to the smallest
+        key is excluded.
 
-        If a max is specified, then output is constrained to values whose
-        keys are less than or equal to the given max, and, if excludemax is
-        specified and true, is further constrained to values whose keys are
-        strictly less than max.  A max value of None is ignored.  If max is
-        None or not specified, and excludemax is true, the value
-        corresponding to the largest key is excluded.
+        If a *max* is specified, then output is constrained to values
+        whose keys are less than or equal to the given *max*, and, if
+        *excludemax* is specified and true, is further constrained to
+        values whose keys are strictly less than *max*. A *max* value of
+        `None` is ignored. If *max* is `None` or not specified, and
+        *excludemax* is true, the value corresponding to the largest key
+        is excluded.
         """
 
     def items(min=None, max=None, excludemin=False, excludemax=False):
-        """Return an :class:`BTrees.Interfaces.IReadSequence` containing the items in the collection.
+        """
+        Return an ``IMinimalSequence`` containing the items in the
+        collection.
 
-        An item is a 2-tuple, a (key, value) pair.
+        An item is a 2-tuple, a ``(key, value)`` pair.
 
-        The type of the :class:`BTrees.Interfaces.IReadSequence` is not specified.  It could be a list
-        or a tuple or some other type.
+        The type of the ``IMinimalSequence`` is not specified. It
+        could be a `list` or a `tuple` or some other type.
 
         All arguments are optional, and may be specified as keyword
         arguments, or by position.
 
-        If a min is specified, then output is constrained to items whose
-        keys are greater than or equal to the given min, and, if excludemin
-        is specified and true, is further constrained to items whose keys
-        are strictly greater than min.  A min value of None is ignored.  If
-        min is None or not specified, and excludemin is true, the item with
-        the smallest key is excluded.
+        If a *min* is specified, then output is constrained to items
+        whose keys are greater than or equal to the given *min*, and,
+        if *excludemin* is specified and true, is further constrained
+        to items whose keys are strictly greater than *min*. A *min*
+        value of `None` is ignored. If *min* is `None` or not
+        specified, and *excludemin* is true, the item with the
+        smallest key is excluded.
 
-        If a max is specified, then output is constrained to items whose
-        keys are less than or equal to the given max, and, if excludemax is
-        specified and true, is further constrained to items whose keys are
-        strictly less than max.  A max value of None is ignored.  If max is
-        None or not specified, and excludemax is true, the item with the
-        largest key is excluded.
+        If a *max* is specified, then output is constrained to items
+        whose keys are less than or equal to the given *max*, and, if
+        *excludemax is specified and true, is further constrained to
+        items whose keys are strictly less than *max*. A *max* value
+        of `None` is ignored. If *max* is `None` or not specified, and
+        *excludemax* is true, the item with the largest key is
+        excluded.
         """
 
 class IDictionaryIsh(IMinimalDictionary):